+++ InventorySolution/Inventory.Core/Domain/Surse/Sursa.cs
@@ -50,7 +50,7 @@ namespace Inventory.Core.Domain
         /// </summary>
         [DataMember]
         [Column(IsDbGenerated = true, IsPrimaryKey = true, Name = "Nr_crt")]
-        public short Id { get; set; } 
+        public int Id { get; set; } 
         
         /// <summary>
         /// Gets or sets the setting identifier
@@ -59,7 +59,7 @@ namespace Inventory.Core.Domain
         [Column(CanBeNull = false, Name = "Sursa")]
         public string Nume { get; set; }
           
-        public override short GetId()
+        public override int GetId()
         {
             return Id;
         }
