+++ InventorySolution/Inventory.Core/Domain/Tipuri/Tip.cs
@@ -48,7 +48,7 @@ namespace Inventory.Core.Domain
         /// </summary>
         [DataMember]
         [Column(IsDbGenerated = true, IsPrimaryKey = true, Name = "Nr_crt")]
-        public short Id { get; set; } 
+        public int Id { get; set; } 
         
         /// <summary>
         /// Gets or sets the setting identifier
@@ -57,7 +57,7 @@ namespace Inventory.Core.Domain
         [Column(CanBeNull = false, Name = "Tip")]
         public string Nume { get; set; }
         
-        public override short GetId()
+        public override int GetId()
         {
             return Id;
         }
